requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.2.0=eigen
absl-py=0.11.0=pyhd3eb1b0_1
alabaster=0.7.12=py36_0
altair=4.1.0=pypi_0
argh=0.26.2=py36_0
argon2-cffi=20.1.0=pypi_0
astor=0.8.1=py36_0
astroid=2.3.3=py36_0
async_generator=1.10=py36h28b3542_0
atomicwrites=1.4.0=py_0
attrs=20.3.0=pyhd3eb1b0_0
autopep8=1.5.4=py_0
babel=2.9.0=pyhd3eb1b0_0
backcall=0.2.0=py_0
base58=2.1.0=pypi_0
bcrypt=3.2.0=py36he774522_0
blas=1.0=mkl
bleach=3.2.1=py_0
blinker=1.4=pypi_0
brotlipy=0.7.0=py36h2bbff1b_1003
ca-certificates=2021.1.19=haa95532_0
cachetools=4.2.0=pypi_0
certifi=2020.12.5=py36haa95532_0
cffi=1.14.4=py36hcd4344a_0
chardet=4.0.0=py36haa95532_1003
click=7.1.2=pypi_0
cloudpickle=1.6.0=py_0
colorama=0.4.4=pyhd3eb1b0_0
cryptography=3.3.1=py36hcd4344a_0
cycler=0.10.0=py36haa95532_0
decorator=4.4.2=py_0
defusedxml=0.6.0=pypi_0
diff-match-patch=20200713=py_0
docutils=0.16=py36_1
entrypoints=0.3=pypi_0
flake8=3.8.4=py_0
freetype=2.10.4=hd328e21_0
future=0.18.2=py36_1
gast=0.2.2=py36_0
gitdb=4.0.5=pypi_0
gitpython=3.1.12=pypi_0
google-pasta=0.2.0=py_0
grpcio=1.31.0=py36he7da953_0
h5py=2.10.0=py36h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pyhd3eb1b0_0
imagesize=1.2.0=py_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
intervaltree=3.1.0=py_0
ipykernel=5.4.3=pypi_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pypi_0
isort=5.6.4=py_0
jedi=0.18.0=pypi_0
jinja2=2.11.2=pyhd3eb1b0_0
joblib=1.0.0=pypi_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py_2
jupyter-client=6.1.11=pypi_0
jupyter_client=6.1.7=py_0
jupyter_core=4.7.0=py36haa95532_0
jupyterlab-widgets=1.0.0=pypi_0
jupyterlab_pygments=0.1.2=py_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_1
keyring=21.8.0=py36haa95532_0
kiwisolver=1.3.0=py36hd77b12b_0
lazy-object-proxy=1.5.2=py36h2bbff1b_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.13.0.1=h200bbdf_0
libsodium=1.0.18=h62dcd97_0
libspatialindex=1.9.3=h33f27b4_0
libtiff=4.1.0=h56a325e_1
lz4-c=1.9.2=hf4a77e7_3
markdown=3.3.3=py36haa95532_0
markupsafe=1.1.1=py36he774522_0
matplotlib=3.3.2=haa95532_0
matplotlib-base=3.3.2=py36hba9282a_0
mccabe=0.6.1=py36_1
mistune=0.8.4=py36he774522_0
mkl=2020.2=256
mkl-service=2.3.0=py36h196d8e1_0
mkl_fft=1.2.0=py36h45dec08_0
mkl_random=1.1.1=py36h47e9c7a_0
nbclient=0.5.1=py_0
nbconvert=6.0.7=py36_0
nbformat=5.1.2=pyhd3eb1b0_1
nest-asyncio=1.4.3=pyhd3eb1b0_0
notebook=6.2.0=pypi_0
numpy=1.19.2=py36hadc3359_0
numpy-base=1.19.2=py36ha3acd2a_0
numpydoc=1.1.0=pyhd3eb1b0_1
olefile=0.46=py36_0
openssl=1.1.1i=h2bbff1b_0
opt_einsum=3.1.0=py_0
packaging=20.8=pyhd3eb1b0_0
pandas=1.1.3=py36ha925a31_0
pandoc=2.11=h9490d1a_0
pandocfilters=1.4.3=py36haa95532_1
paramiko=2.7.2=py_0
parso=0.8.1=pypi_0
pathtools=0.1.2=py_1
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.0=py36h4fa10fc_0
pip=20.3.3=py36haa95532_0
pluggy=0.13.1=py36_0
prometheus-client=0.9.0=pypi_0
prompt-toolkit=3.0.10=pypi_0
protobuf=3.13.0.1=py36ha925a31_1
psutil=5.7.2=py36he774522_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyarrow=2.0.0=pypi_0
pycodestyle=2.6.0=py_0
pycparser=2.20=py_2
pydeck=0.5.0=pypi_0
pydocstyle=5.1.1=py_0
pyflakes=2.2.0=py_0
pygments=2.7.4=pyhd3eb1b0_0
pylint=2.3.1=py36_0
pynacl=1.4.0=py36h62dcd97_1
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py36h6538335_2
pyreadline=2.1=py36_1
pyrsistent=0.17.3=py36he774522_0
pysocks=1.7.1=py36haa95532_0
python=3.6.12=h5500b2f_2
python-dateutil=2.8.1=py_0
python-jsonrpc-server=0.4.0=py_0
python-language-server=0.35.1=py_0
pytz=2020.5=pyhd3eb1b0_0
pywin32=300=pypi_0
pywin32-ctypes=0.2.0=py36_1000
pywinpty=0.5.7=pypi_0
pyyaml=5.3.1=py36he774522_1
pyzmq=21.0.1=pypi_0
qdarkstyle=2.8.1=py_0
qt=5.9.7=vc14h73c81de_0
qtawesome=1.0.1=py_0
qtconsole=4.7.7=py_0
qtpy=1.9.0=py_0
requests=2.25.1=pyhd3eb1b0_0
rope=0.18.0=py_0
rtree=0.9.4=py36h21ff451_1
scikit-learn=0.24.1=pypi_0
scipy=1.5.2=py36h9439919_0
send2trash=1.5.0=pypi_0
setuptools=51.3.3=py36haa95532_4
sip=4.19.8=py36h6538335_0
six=1.15.0=py36haa95532_0
smmap=3.0.4=pypi_0
snowballstemmer=2.0.0=py_0
sortedcontainers=2.3.0=pyhd3eb1b0_0
sphinx=3.4.3=pyhd3eb1b0_0
sphinxcontrib-applehelp=1.0.2=py_0
sphinxcontrib-devhelp=1.0.2=py_0
sphinxcontrib-htmlhelp=1.0.3=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.3=py_0
sphinxcontrib-serializinghtml=1.1.4=py_0
spyder=4.1.5=py36_0
spyder-kernels=1.9.4=py36_0
sqlite=3.33.0=h2a8f88b_0
streamlit=0.74.1=pypi_0
tensorboard=2.0.0=pyhb38c66f_1
tensorflow=2.0.0=eigen_py36h457aea3_0
tensorflow-base=2.0.0=eigen_py36h01553b8_0
tensorflow-estimator=2.0.0=pyh2649769_0
termcolor=1.1.0=py36_1
terminado=0.9.2=pypi_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=he774522_0
toml=0.10.2=pypi_0
toolz=0.11.1=pypi_0
tornado=6.1=py36h2bbff1b_0
traitlets=4.3.3=py36_0
typed-ast=1.4.2=py36h2bbff1b_1
tzlocal=2.1=pypi_0
ujson=4.0.1=py36ha925a31_0
urllib3=1.26.2=pyhd3eb1b0_0
validators=0.18.2=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=1.0.2=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=pypi_0
werkzeug=0.16.1=py_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
win_inet_pton=1.1.0=py36haa95532_0
wincertstore=0.2=py36h7fe50ca_0
wrapt=1.12.1=py36he774522_1
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
yapf=0.30.0=py_0
zeromq=4.3.3=ha925a31_3
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
